refactor: StatePage grid 수정
