+++ frontend/src/components/VaccinationState/VaccinationState.js
@@ -9,11 +9,11 @@ import {
 } from './VaccinationState.styles';
 import { Frame, DonutChart } from '../common';
 
-const VaccinationState = ({ title }) => (
+const VaccinationState = ({ title, withWorld }) => (
   <Container>
     {title && <Title>{title}</Title>}
     <Frame width="100%" showShadow={true}>
-      <FrameContent>
+      <FrameContent withWorld={withWorld}>
         <PrimaryState>
           <DonutChart target="35.8" />
           <Info>
@@ -30,14 +30,16 @@ const VaccinationState = ({ title }) => (
             <div>신규 40,086</div>
           </Info>
         </PrimaryState>
-        <PrimaryState>
-          <DonutChart target="14.6" />
-          <Info>
-            <InfoTitle>세계 완전 접종</InfoTitle>
-            <div>누적 1,096,011,434</div>
-            {/* <div>신규 40,086</div> */}
-          </Info>
-        </PrimaryState>
+        {withWorld && (
+          <PrimaryState>
+            <DonutChart target="14.6" />
+            <Info>
+              <InfoTitle>세계 완전 접종</InfoTitle>
+              <div>누적 1,096,011,434</div>
+              {/* <div>신규 40,086</div> */}
+            </Info>
+          </PrimaryState>
+        )}
       </FrameContent>
     </Frame>
   </Container>
@@ -45,10 +47,12 @@ const VaccinationState = ({ title }) => (
 
 VaccinationState.propTypes = {
   title: PropTypes.string,
+  withWorld: PropTypes.bool,
 };
 
 VaccinationState.defaultProps = {
   title: '',
+  withWorld: true,
 };
 
 export default VaccinationState;
