+++ frontend/src/pages/StatePage/StatePage.styles.js
@@ -7,16 +7,33 @@ const Container = styled.div`
     padding: 3rem 0 0 0;
   }
 `;
-const RegionalContainer = styled.div`
-  margin-top: 2.4rem;
-  width: 100%;
-  display: grid;
-  grid-template-columns: 2fr 1fr;
-  column-gap: 2.4rem;
 
-  @media screen and (max-width: 801px) {
-    grid-template-columns: 1fr;
-    column-gap: 0;
+const Content = styled.div`
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-auto-rows: 15rem;
+  column-gap: 3.2rem;
+  row-gap: 3.2rem;
+
+  & > *:nth-of-type(3n-2) {
+    grid-column-start: 1;
+    grid-column-end: 3;
+    grid-row-start: 1;
+    grid-row-end: 2;
+  }
+
+  & > *:nth-of-type(3n-1) {
+    grid-column-start: 1;
+    grid-column-end: 3;
+    grid-row-start: 2;
+    grid-row-end: 6;
+  }
+
+  & > *:nth-of-type(3n) {
+    grid-column-start: 3;
+    grid-column-end: 4;
+    grid-row-start: 1;
+    grid-row-end: 6;
   }
 `;
 
@@ -29,4 +46,4 @@ const Title = styled.h2`
   }
 `;
 
-export { Container, Title, RegionalContainer };
+export { Container, Content, Title };
